slg: introduce even better way of replacing AOSP apps
* Thanks to the official NikGApps team for the sourced function in __packages__.py [1]
* In nutshell, this way works like this:

>>>
(1). We figure out (com.google.android.dialer) is in the final zip
(2). We read how it's AOSP equivalent app folder could be named
[ By executing getPackageReplacement("com.google.android.dialer") which outputs "Dialer" ]
(3). On module installation, run a script that loops through all system partitions (speicifcally, app and priv-app folders) and searches for that specific directory
(4). If it's found, it will be .replaced by our module, and basically made invisible to the system
<<<

[1] https://github.com/nikgapps/project/blob/main/NikGapps/build/PackageConstants.py

Signed-off-by: chickendrop89 <chickendrop89@gmail.com>